Multiple fixes attempt
